+++ lib/features/history/presentation/providers/history_provider.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:flutter_riverpod/legacy.dart';
 
 import '../../../../services/download_queue.dart';
 import '../../data/history_repository.dart';